Refactored the NAnt scripts to link in the correct framework assemblies for the given target framework.  Several code changes were necessary now that the .NET Compact Framework was actually being compiled.  Several code enhancements had been made that were incompatible with .NET CF 2.0. Fixes [AMQNET-142]. (See https://issues.apache.org/activemq/browse/AMQNET-142)

+++ src/main/csharp/Connection.cs
@@ -46,9 +46,9 @@ namespace Apache.NMS.ActiveMQ
 		private bool connected = false;
 		private bool closed = false;
 		private bool closing = false;
-		private long sessionCounter = 0;
-		private long temporaryDestinationCounter = 0;
-		private long localTransactionCounter;
+		private int sessionCounter = 0;
+		private int temporaryDestinationCounter = 0;
+		private int localTransactionCounter;
 		private readonly AtomicBoolean started = new AtomicBoolean(false);
 		private bool disposed = false;
